requirements.txt
Django==1.7.8
Pillow==2.9.0
djangorestframework==3.1.3